riscv: zero registers when using register-wide operations
what was happening is that instructions like `lb` were only affecting the lower bytes of the register and leaving the top dirty. this would lead to situtations were `cmp_eq` for example was using `xor`, which was failing because of the left-over stuff in the top of the register.

with this commit, we now zero out or truncate depending on the context, to ensure instructions like xor will provide proper results.

+++ src/arch/riscv64/bits.zig
@@ -102,7 +102,7 @@ pub const Memory = struct {
 
 pub const Immediate = union(enum) {
     signed: i32,
-    unsigned: u32,
+    unsigned: u64,
 
     pub fn u(x: u64) Immediate {
         return .{ .unsigned = x };
